some float -> int conversion warnings suppressed (but I wonder if this is really the right thing to do?)
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@21689 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/cocoa/dc.mm
@@ -88,9 +88,9 @@ void wxDC::DoGetTextExtent(const wxString& text, wxCoord *x, wxCoord *y, wxCoord
     NSRange glyphRange = [sm_cocoaNSLayoutManager glyphRangeForTextContainer:sm_cocoaNSTextContainer];
     NSRect usedRect = [sm_cocoaNSLayoutManager usedRectForTextContainer:sm_cocoaNSTextContainer];
     if(x)
-        *x=usedRect.size.width;
+        *x=(int)usedRect.size.width;
     if(y)
-        *y=usedRect.size.height;
+        *y=(int)usedRect.size.height;
     if(descent)
         *descent=0;
     if(externalLeading)
